chore: rename *.state files to lib.ts
