+++ frontend/src/pages/clear/Clear.tsx
@@ -4,12 +4,7 @@ import { CommandCard } from "~/components/common/CommandCard";
 import MettaEditor from "~/components/common/MettaEditor";
 import Loader2 from "lucide-solid/icons/loader-2";
 import { formatedNamespace } from "~/lib/state";
-import {
-  expression,
-  setExpression,
-  isLoading,
-  handleClear,
-} from "./clear.state";
+import { expression, setExpression, isLoading, handleClear } from "./lib";
 
 const ClearPage: Component = () => {
   return (
