+++ frontend/src/pages/export/Export.tsx
@@ -31,7 +31,7 @@ import {
   result,
   exportError,
   handleExport,
-} from "./export.state";
+} from "./lib";
 
 const ExportPage: Component = () => {
   return (
