--- frontend/src/parser/parser.terms.ts
@@ -1,28 +1,0 @@
-// This file was generated by lezer-generator. You probably shouldn't edit it.
-export const LINE_COMMENT = 1,
-  Space = 2,
-  Atom = 3,
-  Symbol = 4,
-  StringLiteral = 5,
-  STRING_LITERAL = 6,
-  BooleanLiteral = 7,
-  BOOLEAN_LITERAL = 8,
-  IntegerLiteral = 9,
-  INTEGER_LITERAL = 10,
-  FloatLiteral = 11,
-  FLOAT_LITERAL = 12,
-  Identifier = 13,
-  IDENTIFIER = 14,
-  SpaceReference = 15,
-  SPACE_REFERENCE = 16,
-  GroundedFunction = 17,
-  GroundedArithmeticFunction = 18,
-  GroundedBooleanFunction = 19,
-  GroundedComparisonFunction = 20,
-  OtherGroundedFunction = 21,
-  GroundedType = 22,
-  CLOSING_PARENTHESIS = 23,
-  OPENING_PARENTHESIS = 24,
-  Expression = 25,
-  Variable = 26,
-  VARIABLE = 27;
